feature: add ingredient information in layout
Now the home_fragment.xml layout contains the information about the
ingredients. This information is stored in a List<Ingredient>.

Ingredient is a new class of the model package. Contains two fields,
name and measure.

Drink now has a getIngredients method that returns a List<Ingredient>.
This methods may fail because we are only checking if the ingredients
are not null, but we ignore the measures. This needs to change.

Besides this, a fab button has been added to obtain a random cocktail.

This commit adds all the new files, modifications and logic for the
changes described.