feat: fix ingredient bug, add new menu and fragment
This commit contains the following:

* Fix to the non-showing ingredients. The issue was that the adapter
wasn't notified about the change in the ingredientList.

Calling to notifyDataChange before setting the new ingredientList
fixes the issue.

* A bottonNavitagionMenu has been added to navigate between fragment
using the NavigationComponent.

Currently there are two fragment, Home and Categories Fragments

* The CategoriesFragment shows the different categories the user can
use to search drinks.

The API to get the information from the alcoholic drinks is added. But
need to be bind to corresponding category. Right now, shows the IDs of
the drinks as a quick test.

* A share button place holder has been added HomeFragment.kt.

This button only shows a toast message. This will change in the future.