test: add testing for Retrofit
There is a new file, RetrofitUnitTest, that contains testing for the
retrofit calls performed in Repository. This calls are:

* getRandomCocktail
* getAlcoholicDrinks

We ensure the name of the drinks is not empty. We assume that the rest
of the information is there. This logic may change in the future.

The commit adds the new mentioned file.