feat: add DrinkListFragment
The app now has a new fragment that presents the different drinks
corresponding to the categories of CategoriesFragment.

Currently, is only possible to navigate to this new fragment by
clicking in the "alcoholic" category. This will change in the future.

The drinks are presented with its image and name in a 3 columns grid.

This commit adds the new files, logic and changes to achieve the
work described above.