feat: data for all categories is shown
The different categories in CategoriesFragment.kt are now functional.

This means that, when the user clicks in one of them, the drinks
corresponding to the category will be se shown in DrinkListFragment.kt.

Besides this, in DrinkAdapter.kt and DrinkListFragment.kt an
onClickListener has been installed. This will be used to open the
drink details when the user clicks on one of them. Right now only
shows the position of the element clicked.

Finally, some code modifications have been added to improve the
re-usability of the code.

This commit add everything necessary for the work described.