chore: refactor test and drink related code
The test code has been refactored and more drink fields are tested.

The code to obtain a drink (random or by id) and a drink list have
changed. Now the Repository and ViewModel returns a Drink or
List<Drink> object instead of a DrinkList object. This makes the code
easier to read.

This commit adds the work described above.