chore: improve getDrink method
The getDrink method in DrinkFragment.kt has been split in two methods:

* getDrink(id: Long = -1) // This method returns an actual drink or
a random one depending on the id

* updateRandomDrink() // This method just updates the drink stored in
the repository

In addition to this, now when a drink is selected via a category, the
only option in the top bar menu is to share it.

This commit adds the changes described above.