feat: add search functionality first step
Now is possible to search drinks by its name. To do so a new menu
in the top bar has been added for the CategoriesFragment.kt.

In order to search, a new GET call has been added to API.kt and
MainController.kt. The same goes for ViewModel.kt.

In the case of Repository.kt, the getDrinkList method calls to
searchDrink() when the query is not one of the categories.

There is still work to do and checks to perform (what if the query
resulting is null?).

In addition to this, some unused files have been deleted.

This commit adds the work described above.