feat: add alert dialog when List<Drink> is empty
An alert dialog will popup when the result of a drink search is empty.

For example:

- The user search a drink using "pepe" as name
- The getDrinkList(query) method of Repository will return and empty
List<Drink). This is added in this commit
- In DrinkListFragment.kt, a check validates that the list is not empty.
Otherwise, the dialog will be shown.
- When the user clicks "ok" in the dialog, it will be returned to the
CategoriesFragment.kt (this is done by simulating a back click).

Finally, some code has been cleaned.

This commit adds the work and logic described above.