chore: add abstract fragment classes, cleaned code
There are two new fragments, DataBindingAbstract and ViewModelAbstract.
This fragments are abstract and the following abstract methods:

DataBindingAbstract:
* abstract fun setupDatabinding(inflater, container)
* abstract fun getRootView(): View

ViewModelAbstract:
* abstract fun getViewModel()

ViewModelAbstract extends from DataBindingAbstract, as all the
fragments will have databinding but not all will use a ViewModel.

The non-abstract fragments inherit from them depending on their
requirements to ensure we have the methods needed.

Besides this, the code for the UI fragments have been cleaned. There
are still work to do regarding the search part of the
CategoriesFragment.kt.

This commit adds the new files, and work described above.