feat: improve search feature
Now, while typing the query the user will see the current result in
the CategoriesFragment.kt. If the query is empty, the categories will
appear.

To do so, this fragment has a RecyclerView whose adapter changes
depending on the query following the above logic.

There's has been a re-structure in the code with a new interface to
handle the click in the different adapters (DrinkAdapter,
CategoriesAdapter). New methods have been added to ViewModel.kt too.

This commit contain what is explained above.