feat: improve search interaction and behavior
Now the search view and the action of searching feels more natural,
like if we were using the Youtube application. For this purposes, new
listener have been added in CategoriesFragment.kt.

The idea is to open the keyboard and search view with just a click,
then once is closed check if the query has results (drinkList not empty).
If so, we set the category adapter. Otherwise, we keep the drink adapter.

In addition to this, the first steps to add the option to search by
ingredient along with name have been added too.

Finally, a missing file (round_outline.xml) for rounding the ImageViews
is also added.

This commit adds everything needed for achieving the work described
above.